Move Surface GPE/lid workaround to new module
Move the MS Surface GPE/lid workaround to a new independent module. This
workaround was previously integrated into the Surface System Aggregator
Module repository, however, this integration does not make much sense as
there is no direct dependency on it and was originally only integrated
as the WSID device was a convenient device to attach a driver to.